+++ grails-app/controllers/org/bbop/apollo/GeneProductController.groovy
@@ -1,9 +1,12 @@
-package org.bbop.apollo.geneProduct
+package org.bbop.apollo
 
 import grails.converters.JSON
 import grails.transaction.Transactional
 import org.bbop.apollo.Feature
+import org.bbop.apollo.GeneProductName
 import org.bbop.apollo.User
+import org.bbop.apollo.geneProduct.GeneProduct
+import org.bbop.apollo.gwt.shared.GlobalPermissionEnum
 import org.bbop.apollo.gwt.shared.PermissionEnum
 import org.bbop.apollo.history.FeatureOperation
 import org.codehaus.groovy.grails.web.json.JSONArray
@@ -15,6 +18,7 @@ import org.restapidoc.pojo.RestApiParamType
 import org.restapidoc.pojo.RestApiVerb
 
 import static org.springframework.http.HttpStatus.NOT_FOUND
+import static org.springframework.http.HttpStatus.UNAUTHORIZED
 
 @Transactional(readOnly = true)
 class GeneProductController {
@@ -27,19 +31,37 @@ class GeneProductController {
 
     @RestApiMethod(description = "Returns a JSON array of all suggested gene product names", path = "/geneProduct/search", verb = RestApiVerb.GET)
     @RestApiParams(params = [
-            @RestApiParam(name = "query", type = "string", paramType = RestApiParamType.QUERY, description = "Query value")
+//        @RestApiParam(name = "featureType", type = "string", paramType = RestApiParamType.QUERY, description = "Feature type")
+        @RestApiParam(name = "organism", type = "string", paramType = RestApiParamType.QUERY, description = "Organism name")
+        , @RestApiParam(name = "query", type = "string", paramType = RestApiParamType.QUERY, description = "Query value")
     ])
     def search() {
         try {
             JSONObject nameJson = permissionService.handleInput(request, params)
-            String query = nameJson.getString("query")
-            JSONArray searchArray = new JSONArray()
-            for(GeneProduct geneProduct in GeneProduct.findAllByProductNameIlike(query+"%")){
-                searchArray.add(geneProduct.productName)
+            log.debug "Showing gene product names ${nameJson}"
+            Organism organism = nameJson.organism ? (Organism.findByCommonName(nameJson.organism) ?: Organism.findById(nameJson.organism)) : null
+
+            List<GeneProductName> geneProductNameList = new ArrayList<>()
+            List<String> geneProductNamesFiltered = new ArrayList<>()
+            geneProductNameList.addAll(GeneProductName.executeQuery("select cc from GeneProductName cc where cc.name like :query", [query: nameJson.query + "%"]))
+            // if there are organism filters for these canned comments for this organism, then apply them
+            // TODO: somehow it is breaking this for organisms
+            List<GeneProductNameOrganismFilter> geneProductNameOrganismFilters = GeneProductNameOrganismFilter.findAllByGeneProductNameInList(geneProductNameList)
+            if (geneProductNameOrganismFilters) {
+                GeneProductNameOrganismFilter.findAllByOrganismAndGeneProductNameInList(organism, geneProductNameList).each {
+                    geneProductNamesFiltered.add(it.geneProductName.name)
+                    geneProductNameList.remove(it.geneProductName)
+                }
             }
-            render searchArray as JSON
+            // otherwise ignore them
+            else {
+                geneProductNameList.each {
+                    geneProductNamesFiltered.add(it.name)
+                }
+            }
+            render geneProductNamesFiltered as JSON
         } catch (Exception e) {
-            def error = [error: 'problem finding gene product names for : '+ e]
+            def error = [error: 'problem finding gene product names for : ' + e]
             log.error(error.error)
             render error as JSON
         }
@@ -47,11 +69,12 @@ class GeneProductController {
 
     @RestApiMethod(description = "Load gene product for feature", path = "/geneProduct", verb = RestApiVerb.POST)
     @RestApiParams(params = [
-            @RestApiParam(name = "username", type = "email", paramType = RestApiParamType.QUERY)
-            , @RestApiParam(name = "password", type = "password", paramType = RestApiParamType.QUERY)
-            , @RestApiParam(name = "uniqueName", type = "Feature uniqueName", paramType = RestApiParamType.QUERY, description = "Gene name to query on")
+        @RestApiParam(name = "username", type = "email", paramType = RestApiParamType.QUERY)
+        , @RestApiParam(name = "password", type = "password", paramType = RestApiParamType.QUERY)
+        , @RestApiParam(name = "uniqueName", type = "Feature uniqueName", paramType = RestApiParamType.QUERY, description = "Gene name to query on")
     ]
     )
+    @Transactional
     def index() {
         JSONObject dataObject = permissionService.handleInput(request, params)
         permissionService.checkPermissions(dataObject, PermissionEnum.READ)
@@ -74,17 +97,17 @@ class GeneProductController {
 //        "references":["refprefix:44444444"]}
     @RestApiMethod(description = "Save New gene product for feature", path = "/geneProduct/save", verb = RestApiVerb.POST)
     @RestApiParams(params = [
-            @RestApiParam(name = "username", type = "email", paramType = RestApiParamType.QUERY)
-            , @RestApiParam(name = "password", type = "password", paramType = RestApiParamType.QUERY)
-            , @RestApiParam(name = "feature", type = "string", paramType = RestApiParamType.QUERY, description = "uniqueName of gene feature to query on")
-            , @RestApiParam(name = "productName", type = "string", paramType = RestApiParamType.QUERY, description = "Name of gene product")
-            , @RestApiParam(name = "alternate", type = "boolean", paramType = RestApiParamType.QUERY, description = "Alternate (default false)")
-            , @RestApiParam(name = "evidenceCode", type = "string", paramType = RestApiParamType.QUERY, description = "Evidence (ECO) CURIE")
-            , @RestApiParam(name = "evidenceCodeLAbel", type = "string", paramType = RestApiParamType.QUERY, description = "Evidence (ECO) Label")
-            , @RestApiParam(name = "negate", type = "boolean", paramType = RestApiParamType.QUERY, description = "Negate evidence (default false)")
-            , @RestApiParam(name = "withOrFrom", type = "string", paramType = RestApiParamType.QUERY, description = "JSON Array of with or from CURIE strings, e.g., {[\"UniProtKB:12312]]\"]}")
-            , @RestApiParam(name = "references", type = "string", paramType = RestApiParamType.QUERY, description = "JSON Array of reference CURIE strings, e.g., {[\"PMID:12312]]\"]}")
-            , @RestApiParam(name = "notes", type = "string", paramType = RestApiParamType.QUERY, description = "JSON Array of reference CURIE strings, e.g., {[\"PMID:12312]]\"]}")
+        @RestApiParam(name = "username", type = "email", paramType = RestApiParamType.QUERY)
+        , @RestApiParam(name = "password", type = "password", paramType = RestApiParamType.QUERY)
+        , @RestApiParam(name = "feature", type = "string", paramType = RestApiParamType.QUERY, description = "uniqueName of gene feature to query on")
+        , @RestApiParam(name = "productName", type = "string", paramType = RestApiParamType.QUERY, description = "Name of gene product")
+        , @RestApiParam(name = "alternate", type = "boolean", paramType = RestApiParamType.QUERY, description = "Alternate (default false)")
+        , @RestApiParam(name = "evidenceCode", type = "string", paramType = RestApiParamType.QUERY, description = "Evidence (ECO) CURIE")
+        , @RestApiParam(name = "evidenceCodeLAbel", type = "string", paramType = RestApiParamType.QUERY, description = "Evidence (ECO) Label")
+        , @RestApiParam(name = "negate", type = "boolean", paramType = RestApiParamType.QUERY, description = "Negate evidence (default false)")
+        , @RestApiParam(name = "withOrFrom", type = "string", paramType = RestApiParamType.QUERY, description = "JSON Array of with or from CURIE strings, e.g., {[\"UniProtKB:12312]]\"]}")
+        , @RestApiParam(name = "references", type = "string", paramType = RestApiParamType.QUERY, description = "JSON Array of reference CURIE strings, e.g., {[\"PMID:12312]]\"]}")
+        , @RestApiParam(name = "notes", type = "string", paramType = RestApiParamType.QUERY, description = "JSON Array of reference CURIE strings, e.g., {[\"PMID:12312]]\"]}")
     ]
     )
     @Transactional
@@ -110,6 +133,7 @@ class GeneProductController {
         geneProduct.addToOwners(user)
         feature.addToGeneProducts(geneProduct)
         geneProduct.save(flush: true, failOnError: true)
+        GeneProductName.findOrSaveByName(geneProduct.productName)
 
         JSONArray oldFeaturesJsonArray = new JSONArray()
         oldFeaturesJsonArray.add(originalFeatureJsonObject)
@@ -118,12 +142,12 @@ class GeneProductController {
         newFeaturesJsonArray.add(currentFeatureJsonObject)
 
         featureEventService.addNewFeatureEvent(FeatureOperation.ADD_GO_ANNOTATION,
-                feature.name,
-                feature.uniqueName,
-                dataObject,
-                oldFeaturesJsonArray,
-                newFeaturesJsonArray,
-                user)
+            feature.name,
+            feature.uniqueName,
+            dataObject,
+            oldFeaturesJsonArray,
+            newFeaturesJsonArray,
+            user)
 
         JSONObject annotations = geneProductService.getAnnotations(feature)
         render annotations as JSON
@@ -131,18 +155,18 @@ class GeneProductController {
 
     @RestApiMethod(description = "Update existing gene products for feature", path = "/geneProduct/update", verb = RestApiVerb.POST)
     @RestApiParams(params = [
-            @RestApiParam(name = "username", type = "email", paramType = RestApiParamType.QUERY)
-            , @RestApiParam(name = "password", type = "password", paramType = RestApiParamType.QUERY)
-            , @RestApiParam(name = "id", type = "string", paramType = RestApiParamType.QUERY, description = "GO Annotation ID to update (required)")
-            , @RestApiParam(name = "feature", type = "string", paramType = RestApiParamType.QUERY, description = "uniqueName of feature to query on")
-            , @RestApiParam(name = "productName", type = "string", paramType = RestApiParamType.QUERY, description = "gene product name")
-            , @RestApiParam(name = "alternate", type = "boolean", paramType = RestApiParamType.QUERY, description = "(default false) alternate")
-            , @RestApiParam(name = "evidenceCode", type = "string", paramType = RestApiParamType.QUERY, description = "Evidence (ECO) CURIE")
-            , @RestApiParam(name = "evidenceCodeLabel", type = "string", paramType = RestApiParamType.QUERY, description = "Evidence (ECO) Label")
-            , @RestApiParam(name = "negate", type = "boolean", paramType = RestApiParamType.QUERY, description = "Negate evidence (default false)")
-            , @RestApiParam(name = "withOrFrom", type = "string", paramType = RestApiParamType.QUERY, description = "JSON Array of with or from CURIE strings, e.g., {[\"UniProtKB:12312\"]}")
-            , @RestApiParam(name = "references", type = "string", paramType = RestApiParamType.QUERY, description = "JSON Array of reference CURIE strings, e.g., {[\"PMID:12312\"]}")
-            , @RestApiParam(name = "notes", type = "string", paramType = RestApiParamType.QUERY, description = "JSON Array of notes strings, e.g., {[\"This is a note\"]}")
+        @RestApiParam(name = "username", type = "email", paramType = RestApiParamType.QUERY)
+        , @RestApiParam(name = "password", type = "password", paramType = RestApiParamType.QUERY)
+        , @RestApiParam(name = "id", type = "string", paramType = RestApiParamType.QUERY, description = "GO Annotation ID to update (required)")
+        , @RestApiParam(name = "feature", type = "string", paramType = RestApiParamType.QUERY, description = "uniqueName of feature to query on")
+        , @RestApiParam(name = "productName", type = "string", paramType = RestApiParamType.QUERY, description = "gene product name")
+        , @RestApiParam(name = "alternate", type = "boolean", paramType = RestApiParamType.QUERY, description = "(default false) alternate")
+        , @RestApiParam(name = "evidenceCode", type = "string", paramType = RestApiParamType.QUERY, description = "Evidence (ECO) CURIE")
+        , @RestApiParam(name = "evidenceCodeLabel", type = "string", paramType = RestApiParamType.QUERY, description = "Evidence (ECO) Label")
+        , @RestApiParam(name = "negate", type = "boolean", paramType = RestApiParamType.QUERY, description = "Negate evidence (default false)")
+        , @RestApiParam(name = "withOrFrom", type = "string", paramType = RestApiParamType.QUERY, description = "JSON Array of with or from CURIE strings, e.g., {[\"UniProtKB:12312\"]}")
+        , @RestApiParam(name = "references", type = "string", paramType = RestApiParamType.QUERY, description = "JSON Array of reference CURIE strings, e.g., {[\"PMID:12312\"]}")
+        , @RestApiParam(name = "notes", type = "string", paramType = RestApiParamType.QUERY, description = "JSON Array of notes strings, e.g., {[\"This is a note\"]}")
     ]
     )
     @Transactional
@@ -167,6 +191,7 @@ class GeneProductController {
         geneProduct.lastUpdated = new Date()
         geneProduct.dateCreated = new Date()
         geneProduct.addToOwners(user)
+        GeneProductName.findOrSaveByName(geneProduct.productName)
         geneProduct.save(flush: true, failOnError: true, insert: false)
 
         JSONArray oldFeaturesJsonArray = new JSONArray()
@@ -176,12 +201,12 @@ class GeneProductController {
         newFeaturesJsonArray.add(currentFeatureJsonObject)
 
         featureEventService.addNewFeatureEvent(FeatureOperation.UPDATE_GO_ANNOTATION,
-                feature.name,
-                feature.uniqueName,
-                dataObject,
-                oldFeaturesJsonArray,
-                newFeaturesJsonArray,
-                user)
+            feature.name,
+            feature.uniqueName,
+            dataObject,
+            oldFeaturesJsonArray,
+            newFeaturesJsonArray,
+            user)
 
         JSONObject annotations = geneProductService.getAnnotations(feature)
         render annotations as JSON
@@ -189,10 +214,10 @@ class GeneProductController {
 
     @RestApiMethod(description = "Delete existing gene product for feature", path = "/geneProduct/delete", verb = RestApiVerb.POST)
     @RestApiParams(params = [
-            @RestApiParam(name = "username", type = "email", paramType = RestApiParamType.QUERY)
-            , @RestApiParam(name = "password", type = "password", paramType = RestApiParamType.QUERY)
-            , @RestApiParam(name = "id", type = "string", paramType = RestApiParamType.QUERY, description = "GO Annotation ID to delete (required)")
-            , @RestApiParam(name = "uniqueName", type = "string", paramType = RestApiParamType.QUERY, description = "Feature uniqueName to remove feature from")
+        @RestApiParam(name = "username", type = "email", paramType = RestApiParamType.QUERY)
+        , @RestApiParam(name = "password", type = "password", paramType = RestApiParamType.QUERY)
+        , @RestApiParam(name = "id", type = "string", paramType = RestApiParamType.QUERY, description = "GO Annotation ID to delete (required)")
+        , @RestApiParam(name = "uniqueName", type = "string", paramType = RestApiParamType.QUERY, description = "Feature uniqueName to remove feature from")
     ]
     )
     @Transactional
@@ -215,14 +240,15 @@ class GeneProductController {
         newFeaturesJsonArray.add(currentFeatureJsonObject)
 
         featureEventService.addNewFeatureEvent(FeatureOperation.REMOVE_GO_ANNOTATION,
-                feature.name,
-                feature.uniqueName,
-                dataObject,
-                oldFeaturesJsonArray,
-                newFeaturesJsonArray,
-                user)
+            feature.name,
+            feature.uniqueName,
+            dataObject,
+            oldFeaturesJsonArray,
+            newFeaturesJsonArray,
+            user)
 
         JSONObject annotations = geneProductService.getAnnotations(feature)
         render annotations as JSON
     }
+
 }